requirements.txt
fastapi
javascript
peewee
pyzpaq
uvicorn
websockets

black
